refactor(unocss): migrate from sub preset to unique config
